feat: begin to implement logic of determining a winner
Instead of playing the game until the board is full, the
loop that simulates a round, will loop until the game is over.

Also, instead of creating a 2-D array with all the possible ways in
which the game is won, to determine a winner, determineWinner(),
checks the rows. Later on, the columns, and the diagonals will be
checked.  The purpose of this approach is to 'divide and conquer'.
.